upload songs with name and duration; change playlist event

+++ src/main/kotlin/com/master/musicroomserver/model/Song.kt
@@ -1,3 +1,3 @@
 package com.master.musicroomserver.model
 
-data class Song(val name: String, val duration: Int)
+data class Song(val name: String, val duration: Long)
